+++ source/Application/Controller/Admin/ThemeMain.php
@@ -12,12 +12,6 @@ use oxRegistry;
 use oxTheme;
 use oxException;
 
-/**
- * Admin article main deliveryset manager.
- * There is possibility to change deliveryset name, article, user
- * and etc.
- * Admin Menu: Shop settings -> Shipping & Handling -> Main Sets.
- */
 class ThemeMain extends \OxidEsales\Eshop\Application\Controller\Admin\AdminDetailsController
 {
     /**