+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - PHP v7 support
 - Composer v1 support
 - Support for NAME constants in Event classes
+- Module re-activation on module settings change
 - Class:
   - `Core\PasswordSaltGenerator`
   - `Internal\Transition\Utility\FallbackTokenGenerator`